requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
playwright>=1.44.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pdfminer.six>=20221105
python-multipart>=0.0.9